Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
strawberry-graphql[fastapi]
httpx